requirements.txt
click==8.1.3
langchain==0.0.135
openai==0.27.2
tiktoken>=0.3.2
python-dotenv
